+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -4,6 +4,9 @@
 
 #include "generated_plugin_registrant.h"
 
+#include <system_theme/system_theme_plugin.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
+  SystemThemePluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("SystemThemePlugin"));
 }