+++ lib/fluent_ui.dart
@@ -33,12 +33,12 @@ export 'src/controls/inputs/toggle_button.dart';
 export 'src/controls/inputs/toggle_switch.dart';
 export 'src/controls/inputs/slider.dart';
 
-export 'src/controls/navigation/pivot.dart';
 export 'src/controls/navigation/navigation_panel/navigation_panel.dart';
 export 'src/controls/navigation/tab_view.dart';
 
 export 'src/controls/surfaces/calendar_view.dart';
 export 'src/controls/surfaces/dialog.dart';
+export 'src/controls/surfaces/flyout/flyout.dart';
 export 'src/controls/surfaces/info_bar.dart';
 export 'src/controls/surfaces/list_tile.dart';
 export 'src/controls/surfaces/progress_indicators.dart';
@@ -47,6 +47,7 @@ export 'src/controls/surfaces/tooltip.dart';
 export 'src/controls/utils/icon.dart';
 export 'src/controls/utils/divider.dart';
 
+export 'src/controls/form/auto_suggest_box.dart';
 export 'src/controls/form/text_box.dart';
 export 'src/controls/form/combo_box.dart';
 export 'src/controls/form/pickers/date_picker.dart';