+++ lib/src/controls/surfaces/list_tile.dart
@@ -1,9 +1,15 @@
 import 'package:fluent_ui/fluent_ui.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 
 const kThreeLineTileHeight = 80.0;
 const kTwoLineTileHeight = 52.0;
 const kOneLineTileHeight = 48.0;
+
+const kDefaultContentPadding = EdgeInsets.symmetric(
+  horizontal: 12.0,
+  vertical: 6.0,
+);
 
 class ListTile extends StatelessWidget {
   const ListTile({
@@ -14,7 +20,7 @@ class ListTile extends StatelessWidget {
     this.title,
     this.subtitle,
     this.isThreeLine = false,
-    this.contentPadding,
+    this.contentPadding = kDefaultContentPadding,
   })  : assert(
           subtitle != null ? title != null : true,
           'To have a subtitle, there must be a title',
@@ -30,9 +36,21 @@ class ListTile extends StatelessWidget {
 
   final bool isThreeLine;
 
-  final EdgeInsetsGeometry? contentPadding;
+  final EdgeInsetsGeometry contentPadding;
 
   bool get isTwoLine => subtitle != null;
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties.add(ColorProperty('tileColor', tileColor));
+    properties.add(FlagProperty('isThreeLine', value: isThreeLine));
+    properties.add(DiagnosticsProperty('shape', shape));
+    properties.add(DiagnosticsProperty<EdgeInsetsGeometry>(
+      'contentPadding',
+      contentPadding,
+    ));
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -48,11 +66,7 @@ class ListTile extends StatelessWidget {
           : isTwoLine
               ? kTwoLineTileHeight
               : kOneLineTileHeight,
-      padding: contentPadding ??
-          EdgeInsets.symmetric(
-            horizontal: 12.0,
-            vertical: 6.0,
-          ),
+      padding: contentPadding,
       child: Row(children: [
         if (leading != null)
           Padding(
@@ -97,7 +111,7 @@ class TappableListTile extends StatelessWidget {
     this.onTap,
     this.focusNode,
     this.autofocus = false,
-    this.contentPadding,
+    this.contentPadding = kDefaultContentPadding,
   }) : super(key: key);
 
   final VoidCallback? onTap;
@@ -114,7 +128,15 @@ class TappableListTile extends StatelessWidget {
   final FocusNode? focusNode;
   final bool autofocus;
 
-  final EdgeInsetsGeometry? contentPadding;
+  final EdgeInsetsGeometry contentPadding;
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties.add(ObjectFlagProperty('onTap', onTap, ifNull: 'disabled'));
+    properties.add(FlagProperty('autofocus', value: autofocus));
+    properties.add(ObjectFlagProperty.has('focusNode', focusNode));
+  }
 
   @override
   Widget build(BuildContext context) {