+++ lib/src/utils/header.dart
@@ -1,4 +1,5 @@
 import 'package:fluent_ui/fluent_ui.dart';
+import 'package:flutter/foundation.dart';
 
 class InfoHeader extends StatelessWidget {
   const InfoHeader({
@@ -11,6 +12,13 @@ class InfoHeader extends StatelessWidget {
   final String header;
   final TextStyle? headerStyle;
   final Widget? child;
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties.add(StringProperty('header', header));
+    properties.add(DiagnosticsProperty('headerStyle', headerStyle));
+  }
 
   @override
   Widget build(BuildContext context) {