+++ lib/src/utils/popup.dart
@@ -13,11 +13,13 @@ class PopUp<T> extends StatefulWidget {
     required this.content,
     required this.contentHeight,
     this.backgroundColor,
+    this.contentWidth,
   }) : super(key: key);
 
   final Widget child;
   final WidgetBuilder content;
   final double contentHeight;
+  final double? contentWidth;
 
   final Color? backgroundColor;
 
@@ -31,21 +33,21 @@ class PopUpState<T> extends State<PopUp> {
   Future<void> openPopup() {
     final NavigatorState navigator = Navigator.of(context);
     final RenderBox itemBox = context.findRenderObject()! as RenderBox;
-    final Rect itemRect = itemBox.localToGlobal(Offset.zero,
-            ancestor: navigator.context.findRenderObject()) &
-        itemBox.size;
+    final Offset target = itemBox.localToGlobal(
+      itemBox.size.center(Offset.zero),
+      ancestor: navigator.context.findRenderObject(),
+    );
+    final Rect itemRect = target & itemBox.size;
     assert(_dropdownRoute == null, 'You can NOT open it twice');
     _dropdownRoute = _PopUpRoute<T>(
+      width: widget.contentWidth,
+      target: target,
       contentHeight: widget.contentHeight,
       content: widget.content(context),
-      buttonRect: EdgeInsets.zero
-          .resolve(Directionality.of(context))
-          .inflateRect(itemRect),
+      buttonRect: itemRect,
       elevation: 4,
       capturedThemes:
           InheritedTheme.capture(from: context, to: navigator.context),
-      dropdownColor: widget.backgroundColor ??
-          context.theme.navigationPanelBackgroundColor,
       transitionAnimationDuration:
           context.theme.mediumAnimationDuration ?? Duration.zero,
     );
@@ -131,23 +133,11 @@ class _PopUpMenuState<T> extends State<_PopUpMenu<T>> {
         scopesRoute: true,
         namesRoute: true,
         explicitChildNodes: true,
-        child: Acrylic(
-          opacity: 0.1,
-          decoration: BoxDecoration(
-            color: widget.dropdownColor ??
-                context.theme.navigationPanelBackgroundColor,
-            borderRadius: BorderRadius.circular(4.0),
-            border: Border.all(
-              color: context.theme.scaffoldBackgroundColor!,
-              width: 0.8,
-            ),
-          ),
-          child: m.Material(
-            type: m.MaterialType.transparency,
-            child: ScrollConfiguration(
-              behavior: const _PopUpScrollBehavior(),
-              child: widget.route.content,
-            ),
+        child: m.Material(
+          type: m.MaterialType.transparency,
+          child: ScrollConfiguration(
+            behavior: const _PopUpScrollBehavior(),
+            child: widget.route.content,
           ),
         ),
       ),
@@ -160,16 +150,23 @@ class _PopUpMenuRouteLayout<T> extends SingleChildLayoutDelegate {
     required this.buttonRect,
     required this.route,
     required this.textDirection,
+    required this.target,
+    required this.verticalOffset,
+    this.width,
   });
 
   final Rect buttonRect;
   final _PopUpRoute<T> route;
   final TextDirection? textDirection;
+  final Offset target;
+  final double verticalOffset;
+  final double? width;
 
   @override
   BoxConstraints getConstraintsForChild(BoxConstraints constraints) {
     final double maxHeight = constraints.maxHeight;
-    final double width = math.min(constraints.maxWidth, buttonRect.width);
+    final double width =
+        this.width ?? math.min(constraints.maxWidth, buttonRect.width);
     return BoxConstraints(
       minWidth: width,
       maxWidth: width,
@@ -180,48 +177,18 @@ class _PopUpMenuRouteLayout<T> extends SingleChildLayoutDelegate {
 
   @override
   Offset getPositionForChild(Size size, Size childSize) {
-    final menuLimits = route.getMenuLimits(buttonRect, size.height);
-    assert(() {
-      final Rect container = Offset.zero & size;
-      if (container.intersect(buttonRect) == buttonRect) {
-        assert(menuLimits.top >= 0.0);
-        assert(menuLimits.top + menuLimits.height <= size.height);
-      }
-      return true;
-    }());
-    assert(textDirection != null);
-    final double left;
-    switch (textDirection!) {
-      case TextDirection.rtl:
-        left = buttonRect.right.clamp(0.0, size.width) - childSize.width;
-        break;
-      case TextDirection.ltr:
-        left = buttonRect.left.clamp(0.0, size.width - childSize.width);
-        break;
-    }
-
-    return Offset(
-      left,
-      menuLimits.top - (childSize.height / 2) + kOneLineTileHeight * 1.25,
+    return positionDependentBox(
+      size: size,
+      childSize: childSize,
+      target: target,
+      preferBelow: false,
+      verticalOffset: 24,
     );
   }
 
   @override
   bool shouldRelayout(_PopUpMenuRouteLayout<T> oldDelegate) {
-    return buttonRect != oldDelegate.buttonRect ||
-        textDirection != oldDelegate.textDirection;
-  }
-}
-
-class _MenuLimits {
-  const _MenuLimits(this.top, this.bottom, this.height);
-  final double top;
-  final double bottom;
-  final double height;
-
-  @override
-  String toString() {
-    return '$top $bottom $height';
+    return target != oldDelegate.target || buttonRect != oldDelegate.buttonRect;
   }
 }
 
@@ -230,11 +197,12 @@ class _PopUpRoute<T> extends PopupRoute<T> {
     required this.content,
     required this.contentHeight,
     required this.buttonRect,
+    required this.target,
     this.elevation = 8,
     required this.capturedThemes,
     required this.transitionAnimationDuration,
     this.barrierLabel,
-    this.dropdownColor,
+    this.width,
   });
 
   final Widget content;
@@ -242,7 +210,8 @@ class _PopUpRoute<T> extends PopupRoute<T> {
   final Rect buttonRect;
   final int elevation;
   final CapturedThemes capturedThemes;
-  final Color? dropdownColor;
+  final Offset target;
+  final double? width;
 
   final Duration transitionAnimationDuration;
 
@@ -259,18 +228,17 @@ class _PopUpRoute<T> extends PopupRoute<T> {
   final String? barrierLabel;
 
   @override
-  Widget buildPage(BuildContext context, Animation<double> animation,
-      Animation<double> secondaryAnimation) {
-    return LayoutBuilder(
-        builder: (BuildContext context, BoxConstraints constraints) {
+  Widget buildPage(context, animation, secondaryAnimation) {
+    return LayoutBuilder(builder: (context, constraints) {
       return _PopUpRoutePage<T>(
+        target: target,
         route: this,
         constraints: constraints,
         content: content,
         buttonRect: buttonRect,
         elevation: elevation,
+        width: width,
         capturedThemes: capturedThemes,
-        dropdownColor: dropdownColor,
       );
     });
   }
@@ -279,32 +247,6 @@ class _PopUpRoute<T> extends PopupRoute<T> {
     if (isActive) {
       navigator?.removeRoute(this);
     }
-  }
-
-  _MenuLimits getMenuLimits(Rect buttonRect, double availableHeight) {
-    final double maxMenuHeight = availableHeight - contentHeight;
-    final double buttonTop = buttonRect.top;
-    final double buttonBottom = math.min(buttonRect.bottom, availableHeight);
-
-    final double topLimit = math.min(kOneLineTileHeight, buttonTop);
-    final double bottomLimit = math.max(availableHeight, buttonBottom);
-
-    double menuTop = (buttonTop) - (buttonRect.height) / 2.0;
-
-    double preferredMenuHeight = EdgeInsets.symmetric(vertical: 8.0).vertical;
-    preferredMenuHeight += contentHeight - availableHeight;
-
-    final double menuHeight = math.min(preferredMenuHeight, maxMenuHeight);
-    double menuBottom = menuTop + menuHeight;
-
-    if (menuTop < topLimit) menuTop = math.min(buttonTop, topLimit);
-
-    if (menuBottom > bottomLimit) {
-      menuBottom = math.max(buttonBottom, bottomLimit);
-      menuTop = menuBottom - menuHeight;
-    }
-
-    return _MenuLimits(menuTop, menuBottom, menuHeight);
   }
 }
 
@@ -315,20 +257,22 @@ class _PopUpRoutePage<T> extends StatelessWidget {
     required this.constraints,
     required this.content,
     required this.buttonRect,
+    required this.target,
     this.elevation = 8,
     required this.capturedThemes,
     this.style,
-    required this.dropdownColor,
+    this.width,
   }) : super(key: key);
 
   final _PopUpRoute<T> route;
   final BoxConstraints constraints;
   final Widget content;
   final Rect buttonRect;
+  final Offset target;
   final int elevation;
   final CapturedThemes capturedThemes;
   final TextStyle? style;
-  final Color? dropdownColor;
+  final double? width;
 
   @override
   Widget build(BuildContext context) {
@@ -339,7 +283,6 @@ class _PopUpRoutePage<T> extends StatelessWidget {
       route: route,
       buttonRect: buttonRect,
       constraints: constraints,
-      dropdownColor: dropdownColor,
     );
 
     return MediaQuery.removePadding(
@@ -352,9 +295,12 @@ class _PopUpRoutePage<T> extends StatelessWidget {
         builder: (BuildContext context) {
           return CustomSingleChildLayout(
             delegate: _PopUpMenuRouteLayout<T>(
+              target: target,
               buttonRect: buttonRect,
               route: route,
               textDirection: textDirection,
+              verticalOffset: 0.0,
+              width: width,
             ),
             child: capturedThemes.wrap(menu),
           );